review-protocol: share rewriteReviewDocumentRuntime with Code OSS and the app host
