+++ apps/review-desktop/code-oss/src/vs/review/common/reviewProtocol.ts
@@ -2174,6 +2174,19 @@ export function parseReviewAgentTraceResponse(
   return parseZod(ReviewAgentTraceResponseSchema, value);
 }
 
+export function rewriteReviewDocumentRuntime(
+  source: string,
+  runtimeUrl: string,
+): string {
+  const runtimeSpecifier = JSON.stringify("review-doc-runtime");
+  if (!source.includes(runtimeSpecifier)) {
+    throw new Error("Review document module has no runtime import.");
+  }
+  return source
+    .split(runtimeSpecifier)
+    .join(JSON.stringify(new URL(runtimeUrl).href));
+}
+
 export function parseZod<T>(
   schema: z.ZodType<T>,
   value: unknown,
