+++ packages/progressive-review/app/src/host/review-client.test.ts
@@ -8,7 +8,6 @@ import {
   reviewFetch,
   reviewStorageKey,
   reviewWasmUrl,
-  rewriteReviewDocumentRuntime,
 } from "./review-client";
 
 const injectedConfig = {
@@ -60,29 +59,6 @@ describe("review host client", () => {
     expect(new Headers(requestInit?.headers).get("x-review-token")).toBe(
       "secret-token",
     );
-  });
-
-  it("rewrites each document rebuild to the current bundled runtime URL", () => {
-    const source =
-      'import { createActiveReviewDocument } from "review-doc-runtime";\nexport const version = 1;';
-    expect(
-      rewriteReviewDocumentRuntime(
-        source,
-        "vscode-file://review/assets/doc-runtime-first.js",
-      ),
-    ).toContain("vscode-file://review/assets/doc-runtime-first.js");
-    expect(
-      rewriteReviewDocumentRuntime(
-        source.replace("version = 1", "version = 2"),
-        "vscode-file://review/assets/doc-runtime-second.js",
-      ),
-    ).toContain("vscode-file://review/assets/doc-runtime-second.js");
-    expect(() =>
-      rewriteReviewDocumentRuntime(
-        "export const version = 3;",
-        "vscode-file://review/assets/doc-runtime.js",
-      ),
-    ).toThrow("no runtime import");
   });
 
   it("retries a rejected document module import", async () => {
