+++ packages/review-protocol/src/index.ts
@@ -151,6 +151,19 @@ export function parseReviewAgentTraceResponse(
   return parseZod(ReviewAgentTraceResponseSchema, value);
 }
 
+export function rewriteReviewDocumentRuntime(
+  source: string,
+  runtimeUrl: string,
+): string {
+  const runtimeSpecifier = JSON.stringify("review-doc-runtime");
+  if (!source.includes(runtimeSpecifier)) {
+    throw new Error("Review document module has no runtime import.");
+  }
+  return source
+    .split(runtimeSpecifier)
+    .join(JSON.stringify(new URL(runtimeUrl).href));
+}
+
 export function parseZod<T>(
   schema: z.ZodType<T>,
   value: unknown,
